requirements.txt
flask==3.0.3
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
opencv-python==4.10.0.84
pillow==10.3.0
numpy==1.26.4